feat: Add default project key (NOT secure)
Add a default project key for WCMC config - since this is shared with anyone who downloads the app, it is not secure. The default key is the sha256 hash of `icca`. It is used to stop Mapeo for ICCA by default syncing with regular Mapeo, but does not have any security guarantees.